Move binding for imports
