+++ cmd/internal/bindings/bindings.go
@@ -1,12 +1,15 @@
-// Copyright 2021 Edward McFarlane. All rights reserved.
+// Copyright 2022 Edward McFarlane. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package starlarkruntimevar
+package bindings
 
 import (
+	_ "gocloud.dev/blob/fileblob"
+	_ "gocloud.dev/blob/memblob"
 	_ "gocloud.dev/runtimevar/blobvar"
 	_ "gocloud.dev/runtimevar/constantvar"
 	_ "gocloud.dev/runtimevar/filevar"
 	_ "gocloud.dev/runtimevar/httpvar"
+	_ "modernc.org/sqlite"
 )
