--- starlarkblob/bindings_aws.go
@@ -1,11 +1,0 @@
-// Copyright 2021 Edward McFarlane. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-// +build aws
-
-package starlarkblob
-
-import (
-	_ "gocloud.dev/blob/s3blob"
-)
