--- starlarksql/bindings.go
@@ -1,12 +1,0 @@
-// Copyright 2021 Edward McFarlane. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-//go:build !js || !wasm
-// +build !js !wasm
-
-package starlarksql
-
-import (
-	_ "modernc.org/sqlite"
-)
