--- starlarksql/bindings_local.go
@@ -1,5 +1,0 @@
-// Copyright 2021 Edward McFarlane. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package starlarksql
